MIP-252 Fixed previous package refactor, throw DefaultMuleException now instead of Unknown
