+++ pom.xml
@@ -103,6 +103,13 @@
             <artifactId>mockserver-client-java</artifactId>
             <version>5.7.0</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.mule.connectors</groupId>
+            <artifactId>mule-http-connector</artifactId>
+            <version>1.5.11</version>
+            <classifier>mule-plugin</classifier>
+        </dependency>
     </dependencies>
     <repositories>
         <repository>
