+++ src/main/java/com/avioconsulting/mule/connector/akv/provider/api/client/AzureClient.java
@@ -1,8 +1,7 @@
-package com.avioconsulting.mule.connector.akv.provider.api.client.client;
+package com.avioconsulting.mule.connector.akv.provider.api.client;
 
 import com.avioconsulting.mule.connector.akv.provider.api.error.AccessDeniedException;
-import com.avioconsulting.mule.connector.akv.provider.api.error.UnknownKeyVaultException;
-import com.avioconsulting.mule.connector.akv.provider.api.client.client.model.OAuthToken;
+import com.avioconsulting.mule.connector.akv.provider.api.client.model.OAuthToken;
 import com.google.gson.Gson;
 import java.io.InputStreamReader;
 import java.io.UnsupportedEncodingException;
@@ -11,6 +10,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+
+import org.mule.runtime.api.exception.DefaultMuleException;
 import org.mule.runtime.http.api.HttpConstants;
 import org.mule.runtime.http.api.client.HttpClient;
 import org.mule.runtime.http.api.client.HttpRequestOptions;
@@ -61,7 +62,7 @@ public class AzureClient {
    * @param timeout         Request timeout in ms, default 30000
    */
   public AzureClient(HttpClient httpClient, String vaultName, String baseUri, String tenantId,
-                     String clientId, String clientSecret, Integer timeout) throws AccessDeniedException {
+                     String clientId, String clientSecret, Integer timeout) throws AccessDeniedException, DefaultMuleException {
     this.httpClient = httpClient;
     this.vaultName = vaultName;
     this.baseUri = baseUri;
@@ -77,7 +78,7 @@ public class AzureClient {
   }
 
 
-  private void authenticate() throws AccessDeniedException {
+  private void authenticate() throws DefaultMuleException {
     Map<String, Object> params = new HashMap<>();
     params.put(PARAM_GRANT_TYPE, GRANT_TYPE_CLIENT_CREDENTIALS);
     params.put(PARAM_CLIENT_ID, clientId);
@@ -108,7 +109,7 @@ public class AzureClient {
       }
     } catch (InterruptedException | ExecutionException | UnsupportedEncodingException e) {
       e.printStackTrace();
-      throw new UnknownKeyVaultException(e.getMessage());
+      throw new DefaultMuleException(e.getMessage());
     }
   }
 
@@ -143,9 +144,13 @@ public class AzureClient {
    */
   public boolean isValid() {
     if (!token.isValid()) {
-      LOGGER.info("Access Token expired, re-authenticating.");
-      authenticate();
-      return token.isValid();
+      try{
+        LOGGER.info("Access Token expired, re-authenticating.");
+        authenticate();
+        return token.isValid();
+      } catch (DefaultMuleException e) {
+        return false;
+      }
     } else {
       return true;
     }
