+++ src/main/java/com/avioconsulting/mule/connector/akv/provider/api/client/AzureKeyVaultClient.java
@@ -1,18 +1,19 @@
-package com.avioconsulting.mule.connector.akv.provider.api.client.client;
+package com.avioconsulting.mule.connector.akv.provider.api.client;
 
 import com.avioconsulting.mule.connector.akv.provider.api.error.CertificateNotFoundException;
 import com.avioconsulting.mule.connector.akv.provider.api.error.KeyNotFoundException;
 import com.avioconsulting.mule.connector.akv.provider.api.error.SecretNotFoundException;
-import com.avioconsulting.mule.connector.akv.provider.api.error.UnknownKeyVaultException;
-import com.avioconsulting.mule.connector.akv.provider.api.client.client.model.Certificate;
-import com.avioconsulting.mule.connector.akv.provider.api.client.client.model.Key;
-import com.avioconsulting.mule.connector.akv.provider.api.client.client.model.KeyVaultError;
-import com.avioconsulting.mule.connector.akv.provider.api.client.client.model.Secret;
+import com.avioconsulting.mule.connector.akv.provider.api.client.model.Certificate;
+import com.avioconsulting.mule.connector.akv.provider.api.client.model.Key;
+import com.avioconsulting.mule.connector.akv.provider.api.client.model.KeyVaultError;
+import com.avioconsulting.mule.connector.akv.provider.api.client.model.Secret;
 import com.google.gson.Gson;
 import java.io.InputStreamReader;
 import java.io.UnsupportedEncodingException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+
+import org.mule.runtime.api.exception.DefaultMuleException;
 import org.mule.runtime.http.api.HttpConstants;
 import org.mule.runtime.http.api.client.HttpClient;
 import org.mule.runtime.http.api.client.HttpRequestOptions;
@@ -33,7 +34,7 @@ public class AzureKeyVaultClient extends AzureClient {
 
   public AzureKeyVaultClient(HttpClient httpClient, String vaultName, String baseUri,
                              String tenantId, String clientId, String clientSecret,
-                             Integer timeout) {
+                             Integer timeout) throws DefaultMuleException {
     super(httpClient, vaultName, baseUri, tenantId, clientId, clientSecret, timeout);
   }
 
@@ -43,7 +44,7 @@ public class AzureKeyVaultClient extends AzureClient {
    * @param secretName    Name of the secret
    * @return Secret       Object containing the secret and associated metadata
    */
-  public Secret getSecret(String secretName) {
+  public Secret getSecret(String secretName) throws DefaultMuleException {
     HttpRequest request = getAuthenticatedHttpRequestBuilder()
         .uri(getHttpBaseUri() + BASE_SECRET_PATH + secretName)
         .addQueryParam(PARAM_API_VERSION, API_VERSION)
@@ -67,13 +68,12 @@ public class AzureKeyVaultClient extends AzureClient {
         if (statusCode == 404) {
           throw new SecretNotFoundException(error.getError().getMessage());
         } else {
-          throw new UnknownKeyVaultException(error.getError().getMessage());
+          throw new DefaultMuleException(error.getError().getMessage());
         }
       }
     } catch (InterruptedException | ExecutionException | UnsupportedEncodingException e) {
       LOGGER.error("Error retrieving secret at " + secretName, e);
-      // rethrow error
-      throw new UnknownKeyVaultException(e.getMessage());
+      throw new DefaultMuleException("Error retrieving secret at " + secretName, e);
     }
   }
 
@@ -83,7 +83,7 @@ public class AzureKeyVaultClient extends AzureClient {
    * @param keyName    Name of the key
    * @return Key       Object containing the key and associated metadata
    */
-  public Key getKey(String keyName) {
+  public Key getKey(String keyName) throws DefaultMuleException {
     HttpRequest request = getAuthenticatedHttpRequestBuilder()
         .uri(getHttpBaseUri() + BASE_KEY_PATH + keyName)
         .addQueryParam(PARAM_API_VERSION, API_VERSION)
@@ -109,12 +109,12 @@ public class AzureKeyVaultClient extends AzureClient {
         if (statusCode == 404) {
           throw new KeyNotFoundException(error.getError().getMessage());
         } else {
-          throw new UnknownKeyVaultException(error.getError().getMessage());
+          throw new DefaultMuleException(error.getError().getMessage());
         }
       }
     } catch (InterruptedException | ExecutionException | UnsupportedEncodingException e) {
       LOGGER.error("Error retrieving Key at " + keyName, e);
-      throw new UnknownKeyVaultException(e.getMessage());
+      throw new DefaultMuleException("Error retrieving Key at " + keyName, e);
     }
   }
 
@@ -124,7 +124,7 @@ public class AzureKeyVaultClient extends AzureClient {
    * @param certificateName    Name of the certificate
    * @return Certificate       Object containing the certificate and associated metadata
    */
-  public Certificate getCertificate(String certificateName) {
+  public Certificate getCertificate(String certificateName) throws DefaultMuleException {
 
     HttpRequest request = getAuthenticatedHttpRequestBuilder()
         .uri(getHttpBaseUri() + BASE_CERTIFICATE_PATH + certificateName + CERTIFICATE_STATUS_PATH)
@@ -152,12 +152,12 @@ public class AzureKeyVaultClient extends AzureClient {
         if (statusCode == 404) {
           throw new CertificateNotFoundException(error.getError().getMessage());
         } else {
-          throw new UnknownKeyVaultException(error.getError().getMessage());
+          throw new DefaultMuleException(error.getError().getMessage());
         }
       }
     } catch (InterruptedException | ExecutionException | UnsupportedEncodingException e) {
       LOGGER.error("Error retrieving Certificate at " + certificateName, e);
-      throw new UnknownKeyVaultException(e.getMessage());
+      throw new DefaultMuleException("Error retrieving certificate at " + certificateName, e);
     }
   }
 }
